refactor: move tests to separate directory
